Refactor BetaLanguageBanner as a separate React Component
